iOS: Add satellite map, plane/ship tracking, fix build, speed up loads
Map overhaul:
- Satellite/hybrid map style instead of standard
- Live aircraft from OpenSky Network (free ADS-B, worldwide)
- Live vessels from openwaters.io AIS (free, worldwide)
- Toggle buttons for events/flights/ships
- Fetch newsAll (9 feeds) instead of newsCore (3) for full category coverage
- Expanded gazetteer (250+ places) with regions, cities, conflict zones
- 400 event annotations + 250 planes + 250 ships, all loaded in parallel

Build fixes (code never compiled before):
- Models.swift: Geo now conforms to Hashable
- RSSParser.swift: Fixed attributeDict scope bug (didEndElement referenced
  a parameter only available in didStartElement, breaking thumbnail fetch)
- APIClient.swift: rssEvents now fetches all feeds concurrently via
  TaskGroup instead of sequentially (biggest load-speed win), capped at
  50 items/feed, centralized geocoding
- AIEngine.swift: Fixed 5 Int/Double millisecond timestamp mismatches
- Added DEVELOPMENT_TEAM to pbxproj for automatic signing

Performance:
- 30s refresh tick now routes only to the visible tab instead of all 7
  models simultaneously, eliminating network contention
- Request timeout reduced 20s -> 12s so one dead feed can't stall startup

+++ ios/WorldIntel/AIEngine.swift
@@ -84,7 +84,7 @@ func clusterEvents(_ events: [WorldEvent]) -> [StoryCluster] {
     }
     return clusters.compactMap { c in
         guard let rep = c.events.max(by: { ($0.severity, $0.published) < ($1.severity, $1.published) }) else { return nil }
-        let now = Date().timeIntervalSince1970 * 1000
+        let now = Int(Date().timeIntervalSince1970 * 1000)
         let lastHour = c.events.filter { $0.published > now - 3_600_000 }.count
         return StoryCluster(
             id: rep.id,
@@ -94,7 +94,7 @@ func clusterEvents(_ events: [WorldEvent]) -> [StoryCluster] {
             count: c.events.count,
             sources: Array(Set(c.events.map(\.source))),
             categories: Array(Set(c.events.map(\.category))),
-            lastSeen: c.events.map(\.published).max() ?? 0,
+            lastSeen: Double(c.events.map(\.published).max() ?? 0),
             momentum: c.events.isEmpty ? 0 : Double(lastHour) / Double(c.events.count)
         )
     }
@@ -112,10 +112,10 @@ struct Spike: Identifiable {
 func detectSpikes(_ events: [WorldEvent], windowCount: Int = 4) -> [Spike] {
     guard events.count >= 8, let newest = events.map(\.published).max(), let oldest = events.map(\.published).min() else { return [] }
     let span = max(newest - oldest, 1)
-    let winMs = span / Double(windowCount)
+    let winMs = Double(span) / Double(windowCount)
     var buckets: [[String: Int]] = Array(repeating: [:], count: windowCount)
     for e in events {
-        let idx = min(windowCount - 1, Int((e.published - oldest) / winMs))
+        let idx = min(windowCount - 1, Int(Double(e.published - oldest) / winMs))
         for t in Set(tokenize(e.title)) {
             buckets[idx][t, default: 0] += 1
         }
@@ -178,7 +178,7 @@ struct Briefing {
 }
 
 func generateBriefing(events: [WorldEvent], hours: Double = 24) -> Briefing {
-    let since = Date().timeIntervalSince1970 * 1000 - hours * 3_600_000
+    let since = Int(Date().timeIntervalSince1970 * 1000 - hours * 3_600_000)
     let recent = events.filter { $0.published >= since }
     let clusters = clusterEvents(recent).prefix(8).map { $0 }
     let breaking = recent.filter { $0.severity >= 4 }.prefix(6).map { $0 }
@@ -259,7 +259,7 @@ struct WorldSummary {
 }
 
 func generateWorldSummary(events: [WorldEvent], hours: Double = 24) -> WorldSummary {
-    let since = Date().timeIntervalSince1970 * 1000 - hours * 3_600_000
+    let since = Int(Date().timeIntervalSince1970 * 1000 - hours * 3_600_000)
     let recent = events.filter { $0.published >= since }
     let clusters = clusterEvents(recent)
     let spikes = detectSpikes(recent)
